move traffic PDE routines into separate file
